Create length & sum.cpp
This code gives the max and minimum.
If you are provide with two numbers and in which the first tell the number of digits the number will have and the second tells the sum of the digits. Then this code will help you to find the minimum  and maximum values of the number which will satisfy the condition other wise will display -1 -1.